fix(GoogleForms): Community Answers hover position
Previously, for questions that requires a hovering Community Answers
value, we were using incorrect screenX and screenY value to determine
the hover position. This causes problems on some clients, preventing
users from choosing an answer.

Now, the correct values clientX and clientY are used, also, hovers
are placed 10px under and to the right, so that users can choose
their answers on Google Forms.